Complete user controller and routes

+++ controllers/user.controller.js
@@ -29,7 +29,7 @@ export const getUserById = async (req, res, next) => {
         }
         res.status(200).json({
             success: true,
-            message: "Users retrieved successfully",
+            message: "User retrieved successfully",
             data: user
         });
     } catch (error) {
@@ -40,6 +40,9 @@ export const getUserById = async (req, res, next) => {
 
 export const updateUser = async (req, res, next) => {
     try {
+        if(req.params.id !== req.user._id.toString()) {
+            return res.status(403).json({ message: 'Forbidden, you can only update your profile' });
+        }
         const user = await User.findByIdAndUpdate(req.params.id, req.body, { new: true }).select('-password');
         if (!user) {
             const error = new Error("User not found");
@@ -58,6 +61,9 @@ export const updateUser = async (req, res, next) => {
 
 export const deleteUser = async (req, res, next) => {
     try {
+        if(req.params.id !== req.user._id.toString()) {
+            return res.status(403).json({ message: 'Unauthorized, you can only delete your profile' });
+        }
         const user = await User.findByIdAndDelete(req.params.id);
         if (!user) {
             const error = new Error("User not found");
